bond-stock-portfolio/requirements.txt
alpaca-trade-api>=3.0.0
PyYAML>=6.0.0
Flask>=3.0.0
